+++ src/main/java/com/minecolonies/coremod/colony/buildings/utils/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains utility classes use by buildings.
+ */
+package com.minecolonies.coremod.colony.buildings.utils;
