+++ src/main/resources/assets/minecolonies/gui/windowhutbuilder.xml
@@ -13,8 +13,9 @@
             <list id="resources" size="78% 85%" pos="30 20">
                 <box size="100% 30" linewidth="2">
                     <label  id="resourceName" size="100 12" pos="5 2" textalign="MIDDLE_LEFT" color="black"/>
-                    <label  id="resourceAvailableNeeded" size="50 12" pos="35 16" textalign="MIDDLE_LEFT" color="black"/>
-                    <button id="resourceAdd" size="16 16" label="↥" pos="0 0"/>
+                    <label  id="resourceMissing" size="50 12" pos="2 16" textalign="MIDDLE_LEFT" color="black"/>
+                    <label  id="resourceAvailableNeeded" size="100% 12" pos="0 16" textalign="MIDDLE" color="black"/>
+                    <button id="resourceAdd" size="16 16" label="↥" pos="0 12"/>
                     <label  id="resourceId" size="1 1" pos="100 100"/>
                     <label  id="resourceQuantity" size="1 1" pos="100 100"/>
                 </box>
